cleanup whitespaces; pre increment iterators; eliminate unneeded creation / copy / destruction of iterator; when you decide to erase on a std:: container, you can pass ( iterator++ )
copy_of_event.times.erase(jt++);


on the erase function instead of

SItimes::iterator kt = jt;
jt++;
copy_of_event.times.erase(kt);

because the result it will be the same.

your version creates a SItimes::iterator, performs a post-increment,
and erases via the saved old iterator value, and last but not least, when
leaving the scope the temp iterator will be destructed.

my version performs a post increment on the iterator and passes the old
value to the erase function of the std::container.
but is incremented and stays valid.

if you perform things like this - pass iterators to a function but afterwards
have them still pointing to a valid entry, use post-increment - otherwise
use pre-increment on iterators. you may have a look at the std:: templates
to see the difference.