properly delete data of CRCInput::RC_WithData - events
about half of the NeutrinoMessages was leaking memory, because either:
* data that was allocated with "char *foo = new char[x]" was deleted
  with "delete foo"
* the type did not match at all

Now (hopefully) all the data gets allocated with "new char[]" and
deleted with "delete[]".
Found with the help of valgrind. Some code still needs to be audited
as it cannot be run under valgrind/memcheck, e.g. the stream2file code.