sectionsd: get rid of timeoutsDMX in CN-Thread
timeoutsDMX was not reset on zap, which could lead to missing
current/next events after switching from a channel without EPG to
one with EPG, which in turn could lead to problems with the parental
pincode check.
Fixing this would be either racy or complicated and since we bail
out after 35 seconds in CN anyway, this check is not needed at all.